requirements.txt
openai
tiktoken
PyYAML==6.0.1
